// src/infrastructure/persistence/generated/prisma/models/User.ts
export type UserModel = runtime.Types.Result.DefaultSelection<Prisma.$UserPayload>
// ...
export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    email: string
    role: $Enums.Role
    password_hash: string
    created_at: Date
  }, ExtArgs["result"]["user"]>
  composites: {}
}